fix all Key tests
